Move PersistenceEvent from core to instrumentations.akka.persistence package
