--- core/src/main/scala/io/scalac/mesmer/core/core.scala
@@ -1,11 +1,0 @@
-package io.scalac.mesmer
-
-import akka.actor.typed.receptionist.ServiceKey
-
-import io.scalac.mesmer.core.event._
-
-package object core {
-
-  val persistenceServiceKey: ServiceKey[PersistenceEvent] =
-    ServiceKey[PersistenceEvent](s"io.scalac.metric.persistence")
-}
