+++ core/src/main/scala/io/scalac/mesmer/core/event/AbstractEvent.scala
@@ -1,0 +1,5 @@
+package io.scalac.mesmer.core.event
+
+trait AbstractEvent extends Any { self =>
+  type Service >: self.type
+}
