+++ core/src/main/scala/io/scalac/mesmer/core/event/Service.scala
@@ -1,7 +1,6 @@
 package io.scalac.mesmer.core.event
 
 import akka.actor.typed.receptionist.ServiceKey
-import io.scalac.mesmer.core._
 
 trait AbstractService {
   type ServiceType
@@ -24,6 +23,4 @@ object Service {
   def apply[T](key: ServiceKey[T]): Service[T] = new Service[T] {
     val serviceKey: ServiceKey[T] = key
   }
-
-  implicit val persistenceService: Service[PersistenceEvent] = Service(persistenceServiceKey)
 }
